% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppenderDigest.R
\name{AppenderDigest}
\alias{AppenderDigest}
\title{Abstract class for digests}
\description{
\strong{Abstract classes} are exported for package developers that want to extend
them, they cannot be instantiated directly.

Abstract class for Appenders that transmit digests of several log events
at once, for example \link{AppenderPushbullet}, \link{AppenderGmail} and
\link{AppenderSendmail}.
}
\section{Fields}{


\describe{
\item{\code{subject_layout}, \code{set_layout(subject_layout)}}{Like \code{layout}, but
used to format the subject/title of the digest. While \code{layout} is applied
to each LogEvent of the digest, \code{subject_layout} is only applied to
the last one.
}
}
}

\section{Methods}{

}

\seealso{
\link{LayoutFormat}, \link{LayoutGlue}

Other abstract classes: 
\code{\link{AppenderMail}}
}
\concept{abstract classes}
