% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppenderDigest.R
\name{AppenderPushbullet}
\alias{AppenderPushbullet}
\title{Send push-notifications via RPushbullet}
\description{
Send push notifications via \href{https://www.pushbullet.com/}{pushbullet}. This
Appender keeps an in-memory buffer like \link{AppenderBuffer}. If the buffer is
flushed, usually because an event of specified magnitude is encountered, all
buffered events are concatenated to a single message that is sent to
\code{\link[RPushbullet:pbPost]{RPushbullet::pbPost()}}. The default behaviour is to push the last 7 log
events in case a \code{fatal} event is encountered.
}
\section{Fields}{


\describe{
\item{\code{apikey}, \code{recipients}, \code{email}, \code{channel}, \code{devices}}{
See \code{\link[RPushbullet:pbPost]{RPushbullet::pbPost()}}
}
}
}

\section{Methods}{

}

\seealso{
\link{LayoutFormat}, \link{LayoutGlue}

Other Appenders: 
\code{\link{AppenderDbi}},
\code{\link{AppenderGmail}},
\code{\link{AppenderRjdbc}},
\code{\link{AppenderSendmail}}
}
\concept{Appenders}
