% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppenderDigest.R
\name{AppenderSendmail}
\alias{AppenderSendmail}
\title{Send emails via sendmailR}
\description{
Send mails via \code{\link[sendmailR:sendmail]{sendmailR::sendmail()}}, which requires that you have access
to an SMTP server that does not require authentication. This
Appender keeps an in-memory buffer like \link{AppenderBuffer}. If the buffer is
flushed, usually because an event of specified magnitude is encountered, all
buffered events are concatenated to a single message. The default behaviour
is to push the last 30 log events in case a \code{fatal} event is encountered.
}
\note{
The default Layout's \code{fmt} indents each log entry with 3 blanks. This
is a workaround so that Microsoft Outlook does not mess up the line breaks.
}
\section{Fields}{


\describe{
\item{\code{headers}, \code{control}}{see \code{\link[sendmailR:sendmail]{sendmailR::sendmail()}}}
}


\describe{
\item{\code{to}, \code{from}, \code{cc}, \code{bcc}}{
\code{character} vectors.
}
\item{\code{html}, \code{set_html()}}{\code{TRUE} or \code{FALSE}. Send a html email message?
This does currently only formats the log contents as monospace verbatim
text.
}
}



\describe{
\item{\code{subject_layout}, \code{set_layout(subject_layout)}}{Like \code{layout}, but
used to format the subject/title of the digest. While \code{layout} is applied
to each LogEvent of the digest, \code{subject_layout} is only applied to
the last one.
}
}
}

\section{Methods}{



}

\examples{
\dontrun{
lgr::AppenderSendmail$new(
  to = "user@ecorp.com",
  control = list(smtpServer = "mail.ecorp.com"),
  from = "lgr_user@yourmail.com"
)
}


}
\seealso{
\link{LayoutFormat}, \link{LayoutGlue}

Other Appenders: 
\code{\link{AppenderDbi}},
\code{\link{AppenderGmail}},
\code{\link{AppenderPushbullet}},
\code{\link{AppenderRjdbc}}
}
\concept{Appenders}
