% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynatraceLayout.R
\name{as_unix_epoch_ms}
\alias{as_unix_epoch_ms}
\title{Convert POSIXct to Unix Epoch milliseconds}
\usage{
as_unix_epoch_ms(x)
}
\arguments{
\item{x}{a \code{POSIXct} vector}
}
\description{
Convert POSIXct to Unix Epoch milliseconds
}
