% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynatraceLayout.R
\docType{data}
\name{DynatraceLayout}
\alias{DynatraceLayout}
\title{A json layout for Dynatrace ingestion}
\format{
An object of class \code{LayoutJson} (inherits from \code{Layout}, \code{R6}) of length 17.
}
\usage{
DynatraceLayout
}
\description{
A json layout for Dynatrace ingestion
}
\keyword{datasets}
