% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayoutElasticSearch.R
\docType{data}
\name{ElasticSearchLayout}
\alias{ElasticSearchLayout}
\title{A json layout for Elasticsearch ingestion}
\format{
An object of class \code{LayoutJson} (inherits from \code{Layout}, \code{R6}) of length 17.
}
\usage{
ElasticSearchLayout
}
\description{
A json layout for Elasticsearch ingestion
}
\keyword{datasets}
