% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayoutElastic.R
\name{LayoutElastic}
\alias{LayoutElastic}
\title{Format log events for output to ElasticSearch}
\value{
The \verb{$new()} method returns an \link[R6:R6Class]{R6::R6} that inherits from
\link[lgr:Layout]{lgr::Layout} and can used as a Layout by an \link[lgr:Appender]{lgr::Appender}.
}
\description{
Format log events for output to ElasticSearch

Format log events for output to ElasticSearch
}
\seealso{
Other Layout: 
\code{\link{LayoutDbi}}

Other database layouts: 
\code{\link{LayoutDbi}}
}
\concept{Layout}
\concept{database layouts}
\section{Super class}{
\code{\link[lgr:Layout]{lgr::Layout}} -> \code{LayoutElastic}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{transform_data}}{a \code{function} to transform the log event to a list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LayoutElastic$new()}}
\item \href{#method-clone}{\code{LayoutElastic$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Layout" data-id="format_event">}\href{../../lgr/html/Layout.html#method-format_event}{\code{lgr::Layout$format_event()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Layout" data-id="toString">}\href{../../lgr/html/Layout.html#method-toString}{\code{lgr::Layout$toString()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayoutElastic$new(
  transform_names = NULL,
  transform_data = function(event) event[["values"]]
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayoutElastic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
