% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SerializerJson.R
\name{Serializer}
\alias{Serializer}
\alias{SerializerJson}
\title{Serializers}
\value{
a \code{Serializer} \link[R6:R6Class]{R6::R6} object for \link{AppenderDbi}.
}
\description{
Serializers are used by \link{AppenderDbi} to store multiple values in a single
text column in a Database table. Usually you just want to use the default
\code{SerializerJson}. Please not that \code{AppenderDbi} as well as \code{Serializers}
are still \strong{experimental}.
}
\examples{
# The defaul Serializer for 'custom fields' columns
SerializerJson$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Serializer-clone}{\code{Serializer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Serializer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Serializer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Serializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[lgrExtra:Serializer]{lgrExtra::Serializer}} -> \code{SerializerJson}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SerializerJson-new}{\code{SerializerJson$new()}}
\item \href{#method-SerializerJson-serialize}{\code{SerializerJson$serialize()}}
\item \href{#method-SerializerJson-clone}{\code{SerializerJson$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SerializerJson-new"></a>}}
\if{latex}{\out{\hypertarget{method-SerializerJson-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SerializerJson$new(
  cols = "*",
  cols_exclude = c("level", "timestamp", "logger", "caller", "msg"),
  col_filter = is.atomic,
  max_nchar = 2048L,
  auto_unbox = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SerializerJson-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-SerializerJson-serialize}{}}}
\subsection{Method \code{serialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SerializerJson$serialize(event)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SerializerJson-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SerializerJson-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SerializerJson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
