% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayoutDynatrace.R
\name{transform_event_dynatrace}
\alias{transform_event_dynatrace}
\title{Transform a log event for Dynatrace}
\usage{
transform_event_dynatrace(event)
}
\arguments{
\item{event}{a \link[lgr:LogEvent]{lgr::LogEvent} object.}
}
\value{
a \code{list} of values that will be serialized to JSON for Dynatrace.
}
\description{
Transform a log event for Dynatrace
}
\seealso{
\url{https://docs.dynatrace.com/docs/discover-dynatrace/references/semantic-dictionary/fields#service}
}
