% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppenderDbi.R
\name{AppenderRjdbc}
\alias{AppenderRjdbc}
\title{Log to databases via RJDBC}
\description{
Log to a database table with the \strong{RJDBC} package. \strong{RJDBC} is only
somewhat  \strong{DBI} compliant and does not work with \link{AppenderDbi}.
\strong{I do not recommend using RJDBC if it can be avoided.}. AppenderRjdbc
is only tested for DB2 databases, and it is likely it will not work properly
for other databases. Please file a bug report if you encounter any issues.
}
\section{Fields}{




Note: \verb{$data} and \code{show()} query the data from the remote database and might
be slow for very large logs.

\describe{
\item{\code{close_on_exit}, \code{set_close_on_exit()}}{\code{TRUE} or \code{FALSE}. Close the
Database connection when the Logger is removed?}
\item{\code{conn}, \code{set_conn(conn)}}{a \link[DBI:dbConnect]{DBI connection}}
\item{\code{table}}{Name of the target database table}
}
}

\section{Methods}{

}

\section{Creating a New Appender}{


An AppenderDbi is linked to a database table via its \code{table} argument. If
the table does not exist it is created either when the Appender is first
instantiated or (more likely) when the first LogEvent would be written to
that table. Rather than to rely on this feature, it is recommended that you
create the target log table first manually using an \verb{SQL CREATE TABLE}
statement as this is safer and more flexible. See also \link{LayoutDbi}.
}

\section{Choosing the Right DBI Layout}{


Layouts for relational database tables are tricky as they have very strict
column types and further restrictions. On top of that implementation details
vary between database backends.

To make setting up \code{AppenderDbi} as painless as possible, the helper function
\code{\link[=select_dbi_layout]{select_dbi_layout()}} tries to automatically determine sensible \link{LayoutDbi}
settings based on \code{conn} and - if it exists in the database already -
\code{table}. If \code{table} does not exist in the database and you start logging, a
new table will be created with the \code{col_types} from \code{layout}.
}

\seealso{
\link{LayoutFormat}, \link{simple_logging}, \link[data.table:data.table]{data.table::data.table}

Other Appenders: 
\code{\link{AppenderDbi}},
\code{\link{AppenderGmail}},
\code{\link{AppenderPushbullet}},
\code{\link{AppenderSendmail}}
}
\concept{Appenders}
