% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynatraceLayout.R
\docType{data}
\name{transform_event_dynatrace}
\alias{transform_event_dynatrace}
\alias{DynatraceLayout}
\title{A json layout for Dynatrace ingestion}
\format{
An object of class \code{LayoutJson} (inherits from \code{Layout}, \code{R6}) of length 17.
}
\usage{
transform_event_dynatrace(event)

DynatraceLayout
}
\arguments{
\item{event}{A \link[lgr:LogEvent]{lgr::LogEvent} object.}
}
\value{
A \code{list} of key-value pairs ready to be serialized to JSON for Dynatrace.
}
\description{
Transforms a \link[lgr:LogEvent]{lgr::LogEvent} object into a list suitable for Dynatracer ingestion.
}
\seealso{
\url{https://docs.dynatrace.com/docs/discover-dynatrace/references/semantic-dictionary/fields#service}
}
\keyword{datasets}
