% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgrExtra-package.R
\docType{package}
\name{lgrExtra-package}
\alias{lgrExtra}
\alias{lgrExtra-package}
\title{lgrExtra: Extra Appenders for 'lgr'}
\description{
Additional appenders for the logging package 'lgr' that support 
  logging to databases, email and push notifications.
}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (\href{https://orcid.org/0000-0003-3344-9851}{ORCID})

}
\keyword{internal}
