% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayoutDynatrace.R
\name{transform_event_dynatrace}
\alias{transform_event_dynatrace}
\title{Transform a log event for Dynatrace}
\usage{
transform_event_dynatrace(event, transform_event_names, excluded_fields)
}
\arguments{
\item{event}{A \link[lgr:LogEvent]{lgr::LogEvent} object.}

\item{transform_event_names}{A named \code{character} vector mapping original
field names to Dynatrace-compatible ones, or a function with a single
mandatory argument that accepts a character vector of field names}

\item{excluded_fields}{A \code{character} vector of field names to exclude from
the final output.}
}
\value{
A \code{list} of key-value pairs ready to be serialized to JSON for Dynatrace.
}
\description{
Converts a \link[lgr:LogEvent]{lgr::LogEvent} object into a list suitable for Dynatrace ingestion,
including optional field renaming and exclusion.
}
\seealso{
\url{https://docs.dynatrace.com/docs/discover-dynatrace/references/semantic-dictionary/fields#service}
}
